Speed up TOC generation
Previously, this theme relied on Sphinx to generate a separate
complete global TOC for every page, and then pruned it to include only
what was required for that page.  This meant that the overall TOC
generation for all N pages took O(N^2) time; when there are several
hundred pages this becomes a significant fraction of the total build
time.

With this change, the theme relies on Sphinx to generate a complete
TOC only once per writer process, and then efficiently computes the
per-page TOCs from that.