Fix OOM by downgrading cssnano
It appears that one of the dependencies of cssnano@5.1.0 (not
specifically listed in `package.json`) causes an out-of-memory crash,
though it is not yet clear which one.

However, due to the way NPM resolves package versions, it is not
sufficient to merely pin the version of `cssnano` itself, as NPM may
still pull in newer versions of the dependencies.

The correct fix would be to identify the problematic cssnano
dependency and add the package with a version constraint to
`package.json`.  Since the package has not yet been identified,
though, instead this commit merely relies on package-lock.json to
avoid pulling in the problematic package.

The package-lock.json from this commit was created by the following
steps:

1. Reverting package-lock.json to its version prior to the commit that
   merges in changes from upstream v8.2.11.
2. Modifying package.json to pin cssnano to 5.0.17 rather than 5.1.0.
3. Running `npm install` to make the minimal updates to the old
   `package-lock.json` to satisfy the new contents of `package.json`.

Note that deleting `package-lock.json` and then recreating it by
running `npm install` will NOT work; it will pull in the problematic
cssnano dependency.