Fix small viewport hamburger menu navigation
Previously, the TOC nodes corresponding to any sections within a page
were also marked as "active" and "current".  The mkdocs-material
hamburger menu was not designed to accommodate more than a single path
from the root of TOC nodes marked active, resulting in the menus for
multiple sections being shown on top of each other at the same time,
leading to confusing and incorrect behavior with the back buttons.

With this change, sections within a page are not marked active or
current, which fixes the hamburger menu.