Suppress warnings for automatic py:param references in signatures
The parameter cross-linking feature of this theme results in
`py:param` references to parameter descriptions being added
automatically to every Python function signature.

When `nitpicky = True`, this leads to a warning for every parameter
that does not have a description, which is not helpful.

This change suppresses those warnings for the implicitly-added
`py:param` references.