Fix interoperability with Esbonio language server (#163) (#258)
The Esbonio language server runs its own asyncio loop context when it
invokes the Sphinx code.  This causes the asyncio.run() call to complain
since it's already being run in the same thread.

Since we're already running a set of threads here using PoolExecutor, we can
run one more thread which runs the asyncio.run() context in a separate
thread.  This mitigates the interoperability issue with the Esbonio
language server at the cost of an additional thread, and at the cost of
blocking the coroutine loop within the Esbonio language server for the
closing of the download set.

(See also swyddfa/esbonio#451)