[docs] reorder palette color options
This a follow up to #256 that orders the colors as they are listed in the CSS bundle (according to hue).

The original ordering was a bit chaotic because it was grabbing the color names from rules specifically for hyperlink elements. But, by only grabbing the color names from rules that define the fg/bg color vars, we get the originally expected results.