Fix the JavaScript search term splitting for compound words
Previously, the JavaScript search code split the search query only on
spaces, which meant that with a search query of a compound word like
"mkdocs-material", for example, would be treated as a single term
"mkdocs-material".  In contrast, when building the search index, the
`/\w+/` regexp is used to split compound words, such that the index
would only contain "mkdoc" and "material" as separate terms.
Consequently, "mkdocs-material" as a single term would never be found.

With this change, the JavaScript search query splitting has been
changed to match what is done when building the index.

Fixes #262.