Make cpp/api_parser extraction of doc comments more robust
For unknown reasons, in some cases libclang fails to include doc
comments in the `Cursor.raw_comment` field when they are present.

As a workaround, don't rely on the `Cursor.raw_comment` field and
instead parse doc comments from the surrounding tokens directly.

This also makes the parsing of the raw comment into doc comments handle
the case of mixed doc comment types.