cpp/apigen: Exclude duplicate UNEXPOSED_DECL entities
libclang seems to generate UNEXPOSED_DECL cursors for
instantiations (not specializations) of variable templates in certain
cases/with certain versions of libclang.

These cursors result in a JSON entity description that is identical that
produced from the cursor corresponding to the original definition. These
are now excluded as duplicates based on the assumption that they will
occur after the "real" cursor.