Store cached data on `app` object rather than `env` object
Previously, the cppreference data and the json schema data were stored
on the `env` object.

Data on the `env` object is pickled and sent back separately with every
chunk of documents that are processed by a worker process. Only
information that relates to the specific documents that have been read
should be stored on the `env` object.