Fix pending_xref issue with format_signatures
Previously, the format_signatures extension could result in pending_xref
nodes with more than one child.  Sphinx assumes that pending_xref nodes
only contain a single child, and if the reference fails to resolve, will
discard all but the first child.

With this change, if there is more than one child to be added to a
pending_xref node, they are wrapped in an `inline` node.

Additionally, extra inserted text (typically whitespace or an adding
")") is no longer added to `pending_xref` nodes when possible.